feat: 모든 자동차 이름을 저장하는 기능 구현
