--- src/main/java/racingcar/Car.java
@@ -1,12 +1,0 @@
-package racingcar;
-
-public class Car {
-    private final String name;
-    private int position = 0;
-
-    public Car(String name) {
-        this.name = name;
-    }
-
-    // 추가 기능 구현
-}
