+++ src/main/java/racingcar/view/OutputView.java
@@ -1,0 +1,4 @@
+package racingcar.view;
+
+public class OutputView {
+}
